Refactor: scaffold slice 9 inbound webhook handlers
