+++ apps/connectshyft-api/src/modules/connectshyft/handlers/postConnectWebhookInbound.ts
@@ -1,0 +1,5 @@
+import { Request, Response } from 'express';
+import { executeConnectShyftInboundWebhookRoute } from '../http/inboundWebhookContext';
+
+export const postConnectWebhookInbound = async (req: Request, res: Response) =>
+  executeConnectShyftInboundWebhookRoute(req, res, 'inbound');
